Rename snackbar service to common naming.
