+++ raw-cms-app/src/modules/shared/components/detail-edit/detail-edit.js
@@ -1,6 +1,6 @@
 import { RawCMS } from '../../../../config/raw-cms.js';
 import { epicSpinners } from '../../../../utils/spinners.js';
-import { snackbarService } from '../../../core/services/snackbar-service.js';
+import { snackbarService } from '../../../core/services/snackbar.service.js';
 import { BaseCrudService } from '../../../shared/services/base-crud-service.js';
 
 const _rawCmsDetailEditEvents = {
